removes the fade in from full width image slice
